+++ packages/terra-application/tests/jest/application-loading-overlay/ApplicationLoadingOverlay.test.jsx
@@ -2,40 +2,26 @@ import React from 'react';
 import { v4 as uuidv4 } from 'uuid';
 
 import ApplicationLoadingOverlay from '../../../src/application-loading-overlay/ApplicationLoadingOverlay';
-import ApplicationLoadingOverlayContext from '../../../src/application-loading-overlay/ApplicationLoadingOverlayContext';
-
-jest.mock('uuid');
 
 describe('ApplicationLoadingOverlay', () => {
-  let reactUseContext;
-  let loadingOverlayContextValue;
+  let mockSpyUuid;
+  let mockSpyContext;
+  const loadingOverlayContextValue = {
+    show: jest.fn(),
+    hide: jest.fn(),
+  };
 
   beforeAll(() => {
-    uuidv4.mockReturnValue('test-id');
-
-    /**
-     * Until Enzyme is updated with full support for hooks, we need to
-     * mock out the useContext implementation.
-     */
-    reactUseContext = React.useContext;
-    React.useContext = (contextValue) => {
-      if (ApplicationLoadingOverlayContext === contextValue) {
-        return loadingOverlayContextValue;
-      }
-      return reactUseContext(contextValue);
-    };
+    mockSpyContext = jest.spyOn(React, 'useContext').mockReturnValue(loadingOverlayContextValue);
+    mockSpyUuid = jest.spyOn(uuidv4, 'v4').mockReturnValue('00000000-0000-0000-0000-000000000000');
   });
 
   afterAll(() => {
-    React.useContext = reactUseContext;
+    mockSpyContext.mockRestore();
+    mockSpyUuid.mockRestore();
   });
 
   it('should render loading overlay as closed', () => {
-    loadingOverlayContextValue = {
-      show: jest.fn(),
-      hide: jest.fn(),
-    };
-
     const wrapper = mount(
       <ApplicationLoadingOverlay />,
     );
@@ -51,11 +37,6 @@ describe('ApplicationLoadingOverlay', () => {
   });
 
   it('should render loading overlay as open', () => {
-    loadingOverlayContextValue = {
-      show: jest.fn(),
-      hide: jest.fn(),
-    };
-
     const wrapper = mount(
       <ApplicationLoadingOverlay isOpen />,
     );
@@ -63,21 +44,16 @@ describe('ApplicationLoadingOverlay', () => {
     expect(wrapper).toMatchSnapshot();
 
     expect(loadingOverlayContextValue.show.mock.calls.length).toBe(1);
-    expect(loadingOverlayContextValue.show.mock.calls[0][0]).toBe('test-id');
+    expect(loadingOverlayContextValue.show.mock.calls[0][0]).toBe('00000000-0000-0000-0000-000000000000');
     expect(loadingOverlayContextValue.hide.mock.calls.length).toBe(0);
 
     wrapper.unmount();
     expect(loadingOverlayContextValue.show.mock.calls.length).toBe(1);
     expect(loadingOverlayContextValue.hide.mock.calls.length).toBe(1);
-    expect(loadingOverlayContextValue.hide.mock.calls[0][0]).toBe('test-id');
+    expect(loadingOverlayContextValue.hide.mock.calls[0][0]).toBe('00000000-0000-0000-0000-000000000000');
   });
 
   it('should transition from open to closed', () => {
-    loadingOverlayContextValue = {
-      show: jest.fn(),
-      hide: jest.fn(),
-    };
-
     const wrapper = mount(
       <ApplicationLoadingOverlay isOpen />,
     );
@@ -85,22 +61,17 @@ describe('ApplicationLoadingOverlay', () => {
     expect(wrapper).toMatchSnapshot();
 
     expect(loadingOverlayContextValue.show.mock.calls.length).toBe(1);
-    expect(loadingOverlayContextValue.show.mock.calls[0][0]).toBe('test-id');
+    expect(loadingOverlayContextValue.show.mock.calls[0][0]).toBe('00000000-0000-0000-0000-000000000000');
     expect(loadingOverlayContextValue.hide.mock.calls.length).toBe(0);
 
     wrapper.setProps({ isOpen: false });
 
     expect(loadingOverlayContextValue.show.mock.calls.length).toBe(1);
     expect(loadingOverlayContextValue.hide.mock.calls.length).toBe(2);
-    expect(loadingOverlayContextValue.hide.mock.calls[0][0]).toBe('test-id');
+    expect(loadingOverlayContextValue.hide.mock.calls[0][0]).toBe('00000000-0000-0000-0000-000000000000');
   });
 
   it('should redisplay loading overlay with new props', () => {
-    loadingOverlayContextValue = {
-      show: jest.fn(),
-      hide: jest.fn(),
-    };
-
     const wrapper = mount(
       <ApplicationLoadingOverlay isOpen />,
     );
@@ -108,23 +79,18 @@ describe('ApplicationLoadingOverlay', () => {
     expect(wrapper).toMatchSnapshot();
 
     expect(loadingOverlayContextValue.show.mock.calls.length).toBe(1);
-    expect(loadingOverlayContextValue.show.mock.calls[0][0]).toBe('test-id');
+    expect(loadingOverlayContextValue.show.mock.calls[0][0]).toBe('00000000-0000-0000-0000-000000000000');
     expect(loadingOverlayContextValue.hide.mock.calls.length).toBe(0);
 
     wrapper.setProps({ backgroundStyle: 'dark' });
 
     expect(loadingOverlayContextValue.show.mock.calls.length).toBe(2);
-    expect(loadingOverlayContextValue.show.mock.calls[1][0]).toBe('test-id');
+    expect(loadingOverlayContextValue.show.mock.calls[1][0]).toBe('00000000-0000-0000-0000-000000000000');
     expect(loadingOverlayContextValue.hide.mock.calls.length).toBe(1);
-    expect(loadingOverlayContextValue.hide.mock.calls[0][0]).toBe('test-id');
+    expect(loadingOverlayContextValue.hide.mock.calls[0][0]).toBe('00000000-0000-0000-0000-000000000000');
   });
 
   it('should honor backgroundStyle prop', () => {
-    loadingOverlayContextValue = {
-      show: jest.fn(),
-      hide: jest.fn(),
-    };
-
     const wrapper = mount(
       <ApplicationLoadingOverlay isOpen backgroundStyle="clear" />,
     );