+++ packages/terra-application/CHANGELOG.md
@@ -6,7 +6,7 @@
   * Updated `getBrowserLocale` method for IE.
   * Updated jest snapshots for terra-icon and terra-button changes.
   * Updated size explanations for ModalManager managed by DisclosureManagerContext.
-  * Updated `uuid` to `8.2.0` for consistency with other components.
+  * Locked `uuid` dependency to `3.4.0`.
 
 * Added
   * Added user action utility button.