+++ packages/terra-application/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.56.0-alpha.0 - (September 8, 2023)
+
 * Changed
   * Updated `getBrowserLocale` method for IE.